rename SimpleTimeSpanFilter to TimeSpanFilter
