+++ demo/sections/stats/index.ts
@@ -2,5 +2,5 @@ import { Section } from '@/demo/router/routeRecords'
 
 export const stats: Section = {
   StatisticKeyValue: () => import('./StatisticKeyValue.vue'),
-  SimpleTimeSpanFilter: () => import('./SimpleTimeSpanFilter.vue'),
+  TimeSpanFilter: () => import('./TimeSpanFilter.vue'),
 }
