+++ src/components/index.ts
@@ -216,7 +216,6 @@ export { default as SchemaPropertyKeyValue } from './SchemaPropertyKeyValue.vue'
 export { default as SearchInput } from './SearchInput.vue'
 export { default as SelectedCount } from './SelectedCount.vue'
 export { default as SeparatedList } from './SeparatedList.vue'
-export { default as SimpleTimeSpanFilter } from './SimpleTimeSpanFilter.vue'
 export { default as StateBadge } from './StateBadge.vue'
 export { default as StateIcon } from './StateIcon.vue'
 export { default as StateListItem } from './StateListItem.vue'
@@ -231,6 +230,7 @@ export { default as TaskRunList } from './TaskRunList.vue'
 export { default as TaskRunListItem } from './TaskRunListItem.vue'
 export { default as TaskRunLogs } from './TaskRunLogs.vue'
 export { default as TaskRunsSort } from './TaskRunsSort.vue'
+export { default as TimeSpanFilter } from './TimeSpanFilter.vue'
 export { default as TimezoneSelect } from './TimezoneSelect.vue'
 export { default as ToastFlowRunCreate } from './ToastFlowRunCreate.vue'
 export { default as VariableCreateModal } from './VariableCreateModal.vue'
